refactor: move class to NumberUtil
