+++ src/main/java/NumberUtil.java
@@ -1,3 +1,14 @@
+import java.util.HashMap;
+
 public class NumberUtil {
+
+    private boolean isInteger(String str) {
+        try {
+            Integer.parseInt(str);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
 
 }
